requirements.txt
blinker==1.4
six==1.12.0
websocket-client==0.56.0
websockets==8.0.2
